requirements.txt
streamlit
PyPDF2
chromadb
sentence-transformers
torch
transformers
python-dotenv